:truck: [#4608] Move useSynchronizeSelect to shared forms package
The hook is actually more broadly usable than in the objects_api, but
for the time being it's only used there.
